sql/planner/optimizer/*/mod.rs -> *.rs
