+++ src/query/service/src/sql/planner/optimizer/property/mod.rs
@@ -14,76 +14,14 @@
 
 mod builder;
 mod enforcer;
+#[allow(clippy::module_inception)]
+mod property;
 mod stat;
 
-use std::collections::HashSet;
-
 pub use builder::RelExpr;
-use common_planner::IndexType;
 pub use enforcer::require_property;
-
-use crate::sql::plans::Scalar;
-
-pub type ColumnSet = HashSet<IndexType>;
-
-#[derive(Default, Clone)]
-pub struct RequiredProperty {
-    pub distribution: Distribution,
-}
-
-impl RequiredProperty {
-    pub fn satisfied_by(&self, physical: &PhysicalProperty) -> bool {
-        self.distribution.satisfied_by(&physical.distribution)
-    }
-}
-
-#[derive(Default, Clone, Debug)]
-pub struct RelationalProperty {
-    pub output_columns: ColumnSet,
-    pub outer_columns: ColumnSet,
-
-    // TODO(leiysky): introduce upper bound of cardinality to
-    // reduce error in estimation.
-    pub cardinality: f64,
-    // We can get the precise row count of a table in databend,
-    // which information is useful to optimize some queries like `COUNT(*)`.
-    pub precise_cardinality: Option<u64>,
-}
-
-#[derive(Default, Clone)]
-pub struct PhysicalProperty {
-    pub distribution: Distribution,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Distribution {
-    Any,
-    Random,
-    Serial,
-    Broadcast,
-    Hash(Vec<Scalar>),
-}
-
-impl Default for Distribution {
-    // Only used for `RequiredProperty`
-    fn default() -> Self {
-        Self::Any
-    }
-}
-
-impl Distribution {
-    /// Check if required distribution is satisfied by given distribution.
-    pub fn satisfied_by(&self, distribution: &Distribution) -> bool {
-        // (required, delivered)
-        match (&self, distribution) {
-            (Distribution::Any, _)
-            | (Distribution::Random, _)
-            | (Distribution::Serial, Distribution::Serial)
-            | (Distribution::Broadcast, Distribution::Broadcast) => true,
-            (Distribution::Hash(ref keys), Distribution::Hash(ref other_keys)) => keys
-                .iter()
-                .all(|key| other_keys.iter().any(|other_key| key == other_key)),
-            _ => false,
-        }
-    }
-}
+pub use property::ColumnSet;
+pub use property::Distribution;
+pub use property::PhysicalProperty;
+pub use property::RelationalProperty;
+pub use property::RequiredProperty;
